Minor bug-fix in port from micro-bunzip: base/limit no longer offset by 1.
The C sources decrement the base/limit pointers before using them, to
avoid wasting a slot (since minlen is always at least 1).  The JavaScript
port increased the initial size of the array (to account for the fact that
it wouldn't be pre-decrementing the pointers), but then *added 1* before
every assignment to the arrays, which is just wrong.  One reasonable option
would be to have avoided the array size increase and then *subtracted 1*
before each indexing operation.  I chose the simpler fix: leave the
increased array size, but kill the "add 1 before indexing" stuff.

It turns out that this bug was harmless, because JavaScript will just
silently increase the size of your array as needed, but it was certainly
misleading (and caused a bug in another project when I substituted
non-resizable typed arrays).